requirements.txt
streamlit
openai
faiss-cpu
spacy
requests
numpy